Fix caching and revalidation issue
