+++ scripts/hooks/useUser.ts
@@ -1,32 +1,11 @@
 "use client";
 
-import { useState, useEffect } from "react";
-import { getCurrentUser } from "../actions/api/auth/auth";
+import { AuthenticationContext } from "@/context/AuthenticationContext";
+import { useContext } from "react";
 
 export function useUser(): User | null {
-  const [user, setUser] = useState<User | null>(null);
-
-  useEffect(() => {
-    async function fetcher() {
-      try {
-        const res = await getCurrentUser();
-        if (!("code" in res.data)) {
-          setUser(res.data as User);
-        } else {
-          setUser(null);
-        }
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      } catch (error) {
-        setUser(null);
-      }
-    }
-
-    fetcher();
-
-    return () => {
-      setUser(null);
-    };
-  }, []); //
+  const ctx = useContext(AuthenticationContext);
+  const user = ctx?.user;
 
   return user;
 }
